+++ src/onegov.election_day/onegov/election_day/utils/media_generator.py
@@ -784,7 +784,7 @@ class MediaGenerator():
         for locale in self.app.locales:
             for item in items:
                 filename = pdf_filename(item, locale)
-                if (self.force or filename not in existing) and item.completed:
+                if filename not in existing and item.completed:
                     path = '{}/{}'.format(self.pdf_dir, filename)
                     if fs.exists(path):
                         fs.remove(path)