+++ src/onegov.election_day/HISTORY.rst
@@ -31,7 +31,7 @@ Changelog
 1.11.0 (2017-06-06)
 ~~~~~~~~~~~~~~~~~~~
 
-- Parses the party of candidates and displays them for majorz elections.
+- Adds PDF signing.
   [msom]
 
 - Parses the party of candidates and displays them for majorz elections.